Create non-functional body code
Adds in the majority of the body code, which can correctly connect and
begin accquiring packets, but which does not function correctly for
actually assembling the final product.  It appears to simply not
assemble multiple bytes into a single file, but we lack clearer
information due to having not yet written tests.  The next commity will
hopefully rectify this.

Co-authored-by: @day00090 <day00090@morris.umn.edu>