Move files to fix tests
In order to fix the nonfunctional automatic testing framework, files
were moved to an internal folder.  PacketFactoryTest had additional
tests added, all of which pass except for the one which handles very
long filenames.

The bats test still does not work, due to seemingly not assembling more
than one packet of a file.  The next step would be to write proper tests
for the FileBuilder class.

Co-authored-by: @day00090 <day00090@morris.umn.edu>